build(js): WIP Attempt switch to semantic-release
The semantic-release tool does determine if the commits since the last
release require a new release, so that's better.  It's also, however, very much an
end-to-end integration and that makes it too rigid for our purposes.  I'd have to do
significant JavaScript customization and that's not worth it for me for now.

Next up, switch back to the Python Commitizen package.

This reverts commit 5e1341686bfd86437e1cc3f78a6794585659427b.