docs(lit): Favor Sphinx for documentation
I haven't used Eleventy/11ty before, but from an hour of reading, it seems like it's
more focused on dynamic behavior, flexible design and such than documentation.  I still
hold that reStructuredText is a superior way to express documentation in source text in
that it approaches the domain with the intention to add *structure* to text that is
readable by both human and machine.  That's as opposed to merely adding markup for
visual presentation.  As such, I haven't been persuaded away from Sphinx for
documentation.