style(js): Enable all linter checks
Check everything until I find reason not to.  [The docs recommend against
this](https://eslint.org/docs/latest/use/configure/configuration-files#using-eslintall)
but since versions are locked/pinned in `./package-lock.json`, we should be able to pin
new failures to specific commits that upgrade ESLint, so I'm not convinced and I want to
try it out.