build(py): Upgrade against the destination branch
Mostly important in CI, the `$ make devel-upgrade-branch` target is intended for use in
an automated upgrade process. As such it always starts against the destination branch,
creates one commit containing the upgrades and *then* pushes the new/reset `*-upgrade`
branch to the remote where the VCS hooks and/or CI will run the tests as appropriate for
a feature branch or merge/pull request.  For purposes of re-using build artifacts and
caches then, it's best if the actual upgrade changes are made and committed on the
destination branch.

Also if a developer wants to perform the upgrades locally, it's annoying to have to
reset the local destination branch every time.  It doesn't matter if the destination
branch is changed under CI since the change to the destination branch won't be pushed,
the only push is of the `*-upgrade` branch.

For both these reasons, make the upgrade changes and commit on the destination
branch, *then* create/reset the `*-upgrade` branch and push if the VCS hooks were
successful for the commit.