build(base): Poorly behaved default image tag
I misunderstood the issue with the Docker `*:latest` default image tag.  Most of the
prose I've read recommending against using it has used the same reproducibility argument
as other arguments for using explicit full versions, for example `rev` in
`./.pre-commit-config.yaml` or OS package versions in Docker images as enforced by `$
hadolint`.  Namely the argument is that build, test, and CI failures should always be
able to be traced to a change in VCS and as such external changes such as version
upgrades should not affect outcomes.  But the issue with `*:latest` is much more
than *only* that:

https://vsupalov.com/docker-latest-tag/