ci(base): Docker socket binding vs build context
When `build(docker): Support docker socket binding` added `CHECKOUT_DIR` to
`./docker-compose*.yml`, it was to ensure that bind volume source paths matched the
paths as seen by the host's `# dockerd` when using Docker socket binding.  I did a
search and replace for `./` which also changed image build context paths but those paths
are from the perspective of the client *sending* the context to the `# dockerd`.
Reverse those specific changes.