docs(base): Reconcile Sphinx PDF vs HTML builders
Before this change the badges were all rendering into one long row in the rendered
`*.pdf`.  The only reason I ever used a `.. list-table::` but without putting each badge
into it's own cell was because I couldn't get rid of the borders in the rendered HTML.
I dug through the default Docutils CSS, the base Sphinx CSS, and the Sphinx theme CSS
and figured out the classes to use to get rid of the borders.  That means I can put each
badge in it's own cell which also resolves the PDF rendering issue.