build(base): Reconcile bind volume worktrees
The previous method for determining `WORKING_DIR_SUFFIX` didn't work where the host `#
dockerd` paths differ from the paths inside the `build-host`, such as in GitHub actions
where the checkout is bind mounted to a different path:

    :5+ docker compose run --rm -T vale --config=./styles/code.ini --ext=.pl
    [41mE100[0m [--config] Runtime error

    path './styles/code.ini' does not exist

    [90m[3mExecution stopped with code 1.[0m[90m[0m
    make[1]: *** [Makefile:709: test-lint-prose-vale-misc] Error 2

After thinking this through, the fundamentals is that the top-level checkouts for the
project needs to be a bind volume inside containers to preserve access to project-wide
data.  For example, a `$ git worktree` creates a `./.gitdir` that references the
top-level checkout so any VCS operations in the worktree still need access to the top
level inside containers.  The working directory inside containers, however, may need to
be a sub-directory, such as a worktree, so we also need the working directory relative
sub-path.  Finally, for other bind volumes, such as the `./var-docker/*/` per-variant
data should *not* be shared between worktrees, so the bind volume source paths *also*
need access to the working directory relative sub-path.