build(py-docker): Avoid image layer rebuilds
The reason for using the `./Makefile` to install tox was DRY but that also means that
this the image layer gets rebuild whenever the `./Makefile` changes and that changes
often and mostly for changes unrelated to the image and that image layer is not trivial
to rebuild.  Finally, it does also make the image a bit mutable with unfixed versions.