build(py-docker): Fix pull other py version images
In CI, the jobs for non-default Python versions, such as ``debian-py312``, was pulling
the default ``debian-py311`` image to compile the requirements for ``$ tox -e "build"``.
The right fix for that is to only compile those requirements outside any image, directly
in the build host.

That still leaves the issue of how to compile those requirements for all Python versions
as a check that all changes to `./requirements/build.txt.in` are reflected in the
compiled versions.  Cover that by adding the targets for compiling those versions to the
`test-docker-devel-%` targets.

Finally, to avoid compiling these requirements redundantly on multiple jobs, remove the
`./requirements/build.txt.in` prerequisite from the build host version of the target so
that it's only a prerequisite for the Docker version of the target.