build(base): Lock/freeze devel tool image digests
There are 2 important concerns in managing the versions of dependencies: keeping
up-to-date with new releases and reproducible deployments.

The latter is important for reliability, yes, but it's more important, IMO,
for *accountability*. The greatest CI sin, IMO, is false failures unrelated to the
changes being tested. Any new failure should be accountable to a specific change in VCS
and as such if a failure is caused by a tool upgrade, that failure should be traceable
to the commit that upgraded that tool.

Keeping up-to-date with new releases is, however, just as important IMO and in my
experience it seems to get short shrift. Upgrading dependencies is import for security,
sure, but it's also important for maintainability. It becomes harder and harder to keep
up to date the more out of date dependencies become. It's also important for code
quality and best practices because the kind of things maintainers learn as they make
changes necessary to keep up to date are a part of the ongoing education that makes them
better developers.

These 2 concerns, then, are equally important. More than that, I think we in the
software world get it wrong when we think of them as separate concerns at all. IMO, they
should be thought of as 2 parts of one software best practice. Call it dependency
version hygiene.

CI pulled a newer container image of a linter today that resulted in a false failure
unrelated to the changes pushed, the horror! For some time now I've been looking for a
good solution to this for projects using Docker. Unfortunately, the method for saying
which container images to track upstream, `image: "foo/bar:latest-debian"`, is also the
same method for locking versions, `image: "foo/bar@sha256:*"`. I really think this is a
problem that Docker should find a way to solve elegantly for `./docker-compose*.yml`
projects, but I have yet to find one. So it's time to give up on the perfect and come up
with something good enough for now.

Because Docker doesn't provide a way to express both what to track and what the current
digest is in our `./docker-compose*.yml` files, we need something else to hold one of
those two pieces of information. This approach exploits the fact that `$ docker
compose *` always reads variables from `./.env` and substitutes them in
`./docker-compose*.yml` files. Specify the tag to track upstream updates from as the
default for variable and then use the value of that variable in `./.env` to lock the
specific image version digest.