ci(base): Reconcile Docker base image/socket group
Previous attempt didn't work because we can't add a GID to the users list of
supplimental groups, only an actual group name. The base Docker image, however, adds a
`docker` group, presumably when it installs the Docker packages. Understanding all that
now, I think the "most correct" solution is to ensure the Docker socket's group actually
has a named group and if we have to add that group, that we call it `docker` which means
removing that group until needed.

From GitHub Actions:

    /usr/local/bin/init-job.sh:61+adduser runner UNKNOWN
    addgroup: unknown group UNKNOWN
    Error: Process completed with exit code 1.